added: Reduce (#106)
Reduce continually applies the provided function over the slice. Reducing the elements to a single value.

Returns a zero value of ElementType if there are no elements in the slice. It will panic if the reducer is nil and the slice has more than one element (required to invoke reduce).

Otherwise returns a ​result of applying reducer from left to right.

Fixes #87